@@ -22,23 +22,23 @@
     "@skeletonlabs/tw-plugin": "0.4.0",
     "@sveltejs/adapter-auto": "^3.0.0",
     "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/enhanced-img": "^0.2.0",
+    "@sveltejs/enhanced-img": "^0.3.9",
     "@sveltejs/kit": "2.21.1",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tauri-apps/cli": "^1.5.12",
     "@types/node": "20.12.7",
     "autoprefixer": "10.4.19",
     "postcss": "8.4.38",
     "sass": "^1.75.0",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.6.0",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "svelte-loading-spinners": "^0.3.6",
-    "svelte-preprocess": "^5.1.4",
+    "svelte-preprocess": "^6.0.0",
     "svelte-range-slider-pips": "^3.2.2",
     "tailwindcss": "3.4.3",
     "tslib": "^2.4.1",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.3",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4",
     "vite-plugin-tailwind-purgecss": "0.3.2",
     "vitest": "^1.0.0"
   },
