@@ -1,8 +1,12 @@
 <script lang="ts">
     import { propControlWord1 } from '$lib/stores/state';
 
-    export let index: number = 0;
-    let value = $propControlWord1 >> index & 1;
+    interface Props {
+        index?: number;
+    }
+
+    let { index = 0 }: Props = $props();
+    let value = $state($propControlWord1 >> index & 1);
 
     let changeValue = () => {
         $propControlWord1 = $propControlWord1 ^ (1 << index);
@@ -12,7 +16,7 @@
 </script>
 
 <div class="flex flex-col items-center">
-    <button on:click={changeValue}>▲</button>
+    <button onclick={changeValue}>▲</button>
     <input type="text" maxlength="1" class="text-center input rounded-md " bind:value={value}>
-    <button on:click={changeValue}>▼</button>
+    <button onclick={changeValue}>▼</button>
 </div>
