@@ -1,14 +1,21 @@
 <script lang="ts">
+    import { run } from 'svelte/legacy';
+
     import { onDestroy } from 'svelte';
     import { GreenHVALTurnedOn, RedHVALTurnedOn } from '$lib/stores/state';
 
-    // True means the light will be green while false is red.
-    export let innerClass: string = "";
-    export let isGreen: boolean = true;
+    
+    interface Props {
+        // True means the light will be green while false is red.
+        innerClass?: string;
+        isGreen?: boolean;
+    }
+
+    let { innerClass = "", isGreen = true }: Props = $props();
 
     const greenShadow = 'shadow-[0_0_10px_rgba(0,255,0,0.8),0_0_20px_rgba(0,255,0,0.6),0_0_30px_rgba(0,255,0,0.4),0_0_40px_rgba(0,255,0,0.2)]';
     const redShadow = 'shadow-[0_0_10px_rgba(255,0,0,0.8),0_0_20px_rgba(255,0,0,0.6),0_0_30px_rgba(255,0,0,0.4),0_0_40px_rgba(255,0,0,0.2)]';
-    let shadow = "";
+    let shadow = $state("");
 
     // red is off or flashing
     // green is on or off
@@ -16,11 +23,11 @@
     const green: string = "bg-[#0f0]";
     const red: string = "bg-[#f00]";
     const colorOff = "bg-surface-800";
-    let color: string = "bg-[#000]";
+    let color: string = $state("bg-[#000]");
 
-    let blinkOn: boolean = false;
+    let blinkOn: boolean = $state(false);
     let timer: number = 200;
-    let blinkInterval: NodeJS.Timeout | null = null;
+    let blinkInterval: NodeJS.Timeout | null = $state(null);
 
     function startBlinking() {
         if (blinkInterval) {
@@ -33,7 +40,7 @@
         }, timer);
     }
 
-    $: {
+    run(() => {
         if ($RedHVALTurnedOn && !isGreen) {
             startBlinking();
         } else if (!isGreen) {
@@ -51,7 +58,7 @@
             color = colorOff;
             shadow = colorOff;
         }
-    }
+    });
 
     onDestroy(() => {
         if (blinkInterval) {
@@ -66,4 +73,4 @@
     rounded-full {blinkOn ? color : 'bg-surface-800'} size-7
     {shadow}
     "
-/>
+></div>
