@@ -2,14 +2,26 @@
     import {invoke} from '@tauri-apps/api/tauri';
     import {EventChannel, type NamedCommand, util} from "$lib";
 
-    export let offCmd: NamedCommand;
-    export let onCmd: NamedCommand;
-    export let val: number = 0;
-    export let callback: (val:number) => void = () => {};
-    export let disabled: boolean = false;
 
-    // for binding
-    export let status: boolean = false;
+    
+    interface Props {
+        offCmd: NamedCommand;
+        onCmd: NamedCommand;
+        val?: number;
+        callback?: (val:number) => void;
+        disabled?: boolean;
+        // for binding
+        status?: boolean;
+    }
+
+    let {
+        offCmd,
+        onCmd,
+        val = 0,
+        callback = () => {},
+        disabled = false,
+        status = $bindable(false)
+    }: Props = $props();
 
     const toggleOff = () => {
         invoke('send_command', {cmdName:offCmd, val}).then((r) => {
@@ -39,13 +51,13 @@
 <div class="flex">
     <button class="btn rounded-none rounded-l-lg font-number py-2 text-surface-50 bg-surface-700"
             class:active={!status}
-            on:click={toggleOff}
+            onclick={toggleOff}
             disabled={disabled}>
         Off
     </button>
     <button class="btn rounded-none rounded-r-lg font-number py-2 text-surface-50 bg-surface-700"
             class:active={status}
-            on:click={toggleOn}
+            onclick={toggleOn}
             disabled={disabled}>
         On
     </button>
