@@ -3,9 +3,14 @@
     import { slide } from 'svelte/transition';
     import { ArrowUp } from 'carbon-icons-svelte';
 
-    export let title: string;
+    interface Props {
+        title: string;
+        content?: import('svelte').Snippet;
+    }
 
-    let visible: boolean = true;
+    let { title, content }: Props = $props();
+
+    let visible: boolean = $state(true);
 
     let toggleVisibility = () => {
         visible = !visible;
@@ -15,12 +20,12 @@
 <Tile containerClass="col-span-full">
     <div class="flex justify-between">
         <h3 class="text-xl font-normal">
-            <button on:click={toggleVisibility}>
+            <button onclick={toggleVisibility}>
                 {title}
             </button>
         </h3>
         <div class="transition-transform duration-300 flex items-center text-center " class:rotate-180={!visible}>
-            <button on:click={toggleVisibility}>
+            <button onclick={toggleVisibility}>
                 <ArrowUp size={16}/>
             </button>
         </div>
@@ -28,7 +33,7 @@
     {#if visible}
         <div transition:slide={{ duration: 300 }}>
             <hr class="col-span-full my-4">
-            <slot name="content"/>
+            {@render content?.()}
         </div>
     {/if}
 </Tile>
