@@ -1,9 +1,19 @@
 <script lang="ts">
-    export let columns:string;
-    export let rows:string;
-    export let className:string = '';
+    interface Props {
+        columns: string;
+        rows: string;
+        className?: string;
+        children?: import('svelte').Snippet;
+    }
+
+    let {
+        columns,
+        rows,
+        className = '',
+        children
+    }: Props = $props();
 </script>
 
 <div class="grid gap-2 {className}" style="grid-template-columns: {columns}; grid-template-rows: {rows};">
-    <slot />
+    {@render children?.()}
 </div>
