@@ -10,7 +10,7 @@
 <div class="flex flex-col gap-4 p-5 h-[95.5vh] flex-grow">
     <div class="flex flex-row w-full justify-end">
         <button
-            on:click={() => {menuOpen.set(false)}}
+            onclick={() => {menuOpen.set(false)}}
         >
             <Close size={20}/>
         </button>
@@ -19,12 +19,12 @@
     <span class="text-primary-500 text-3xl mt-5 p-2 mr-16">
         Delft Hyperloop
     </span>
-    <div class="w-full border-primary-500" style="border-width: 0.5px;"/>
+    <div class="w-full border-primary-500" style="border-width: 0.5px;"></div>
 
     {#if !$debugModeActive}
         <button
             class={btnClass}
-            on:click={() => {debugModeActive.set(true)}}
+            onclick={() => {debugModeActive.set(true)}}
         >
             <Icon icon="mdi:bug-outline" class="w-6 h-6"/>
             <span>
@@ -34,7 +34,7 @@
     {:else}
         <button
             class={btnClass}
-            on:click={() => {debugModeActive.set(false)}}
+            onclick={() => {debugModeActive.set(false)}}
         >
             <Icon icon="mdi:bug-outline" class="w-6 h-6"/>
             <span>
@@ -44,14 +44,14 @@
     {/if}
     <button
         class={btnClass}
-        on:click={() => new ViewWindow("Charts", `/view/charts`)}
+        onclick={() => new ViewWindow("Charts", `/view/charts`)}
     >
         <ChartLineSmooth size={20}/>
         Graph Visualizer
     </button>
     <button
         class={btnClass}
-        on:click={() => {threeDModeActive.set(false)}}
+        onclick={() => {threeDModeActive.set(false)}}
     >
         <WatsonHealth3DMprToggle size={20}/>
         <span>
