@@ -19,7 +19,11 @@
         window.removeEventListener('resize', updateHeight);
     });
 
-    export let data: PageData;
+    interface Props {
+        data: PageData;
+    }
+
+    let { data }: Props = $props();
 </script>
 
 <div class="h-full">
